JIT tests cleanup: * move meta AB tests to test_covidcast_meta from test_covidcast_meta_cache_updater * move test data files to testdata directory
