+++ testdata/acquisition/covidcast/test-jhu-data-small.csv
