+++ testdata/acquisition/covidcast/test-jhu-data-smaller.csv
